requirements.txt
fastapi
uvicorn
aiohttp
beautifulsoup4
newspaper3k
markdownify
